requirements.txt
streamlit
sqlalchemy
psycopg2-binary
pandas
pdfminer.six
beautifulsoup4
requests
python-dotenv
thefuzz